requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
sqlparse==0.5.5
